core: frontend: Move stream delete method to store
